fix: merge from main, requestedquantity fix

+++ wmsBack/src/main/java/com/isd/wms/dto/replenishment/ReplenishmentTaskCreateRequest.java
@@ -2,7 +2,7 @@ package com.isd.wms.dto.replenishment;
 
 public record ReplenishmentTaskCreateRequest(
         Long productId,
-        Long requestedQuantity,
+        Integer requestedQuantity,
         Long sourceLocationId,
         Long destinationLocationId
 ) {
