+++ wmsBack/src/main/java/com/isd/wms/dto/replenishment/ReplenishmentTaskUpdateRequest.java
@@ -5,7 +5,7 @@ import com.isd.wms.enums.ReplenishmentTaskStatus;
 public record ReplenishmentTaskUpdateRequest(
         Long productId,
         Long operatorId,
-        Long requestedQuantity,
+        Integer requestedQuantity,
         ReplenishmentTaskStatus status,
         Long sourceLocationId,
         Long destinationLocationId
