+++ wmsBack/src/main/java/com/isd/wms/service/ReplenishmentTaskService.java
@@ -178,7 +178,7 @@ public class ReplenishmentTaskService {
     private void validateReplenishmentTaskRequest(
             Long productId,
             Long operatorId,
-            Long requestedQuantity,
+            Integer requestedQuantity,
             ReplenishmentTaskStatus status,
             Long sourceLocationId,
             Long destinationLocationId) {
